OCR/requirements.txt
Flask
Pillow
paddlepaddle
paddleocr